create store for recipe area and move keyboard navigation logic
