+++ components/examine/recipe/tabs/index.vue
@@ -38,7 +38,7 @@
 
 <script setup lang="ts">
 import { TabGroup, TabPanels, TabPanel } from '@headlessui/vue'
-import { useExamineTabs } from '~/store/examine/tabs';
+import { useExamineRecipe } from '~/store/examine/recipe';
 
 const conflicts = ref<InstanceType<typeof TabPanel>>()
 const conditions = ref<InstanceType<typeof TabPanel>>()
@@ -67,7 +67,7 @@ function setTop(elem: TabElementRef, top: number) {
 }
 
 function onChangeTab(index: number) {
-  useExamineTabs().selectedTabIndex = index
+  useExamineRecipe().selectedTabIndex = index
   setTop(panels, scrollPositions[index])
   const panelTop = getTop(panels)
   scrollPositions[lastIndex.value] = panelTop - getTop(tabs[lastIndex.value])
