--- store/examine/tabs.ts
@@ -1,4 +1,0 @@
-export const useExamineTabs = defineStore('examine-tabs', () => {
-  const selectedTabIndex = ref(0)
-  return { selectedTabIndex }
-})
